revert absolute build out debug system
